changed type to size_t

+++ hachaturyanov.ivan/S1/printresults.hpp
@@ -5,7 +5,7 @@
 #include "list.hpp"
 
 namespace hachaturyanov {
-  using Pair = std::pair< std::string, hachaturyanov::List< int > >;
+  using Pair = std::pair< std::string, hachaturyanov::List< size_t > >;
 
   void printNames(std::ostream &out, const List< Pair > &pairs);
 }
