+++ hachaturyanov.ivan/S1/readpairs.hpp
@@ -5,7 +5,7 @@
 #include "list.hpp"
 
 namespace hachaturyanov {
-  using Pair = std::pair< std::string, hachaturyanov::List< int > >;
+  using Pair = std::pair< std::string, hachaturyanov::List< size_t > >;
 
   void skipSpaces(std::istream &in);
   bool isLineEnded(std::istream &in);
